+++ src/components/Navbar/Navbar.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import styled from 'styled-components';
 import { useNavigate } from 'react-router-dom';
 import menu from '../../assets/images/bag-outline.png';
-import { useSignOut } from '../../hooks/queries/useSignOut';
+import { useSignOut } from '../../hooks/queries/Auth/useSignOut';
 
 const NavLayout = styled.header`
   width: 100%;
