+++ src/hooks/queries/Auth/useSignIn.jsx
@@ -1,6 +1,6 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { useNavigate } from 'react-router-dom';
-import authAPI from '../../apis/authAPI';
+import authAPI from '../../../apis/authAPI';
 
 const useSignIn = () => {
   const queryClient = useQueryClient();
